# src/apiserver/openapi_server/models/mec_creation.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server.models.geolocation_tile_list import GeolocationTileList  # noqa: F401,E501
from openapi_server.models.mec_creation_resources import MECCreationResources  # noqa: F401,E501
from openapi_server.models.sb_service import SBService  # noqa: F401,E501
from openapi_server import util


class MECCreation(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, name: str=None, lat: str=None, lng: str=None, organization: str=None, resources: MECCreationResources=None, sb_services: List[SBService]=None, props: object=None, geolocation: GeolocationTileList=None):  # noqa: E501
        """MECCreation - a model defined in Swagger

        :param name: The name of this MECCreation.  # noqa: E501
        :type name: str
        :param lat: The lat of this MECCreation.  # noqa: E501
        :type lat: str
        :param lng: The lng of this MECCreation.  # noqa: E501
        :type lng: str
        :param organization: The organization of this MECCreation.  # noqa: E501
        :type organization: str
        :param resources: The resources of this MECCreation.  # noqa: E501
        :type resources: MECCreationResources
        :param sb_services: The sb_services of this MECCreation.  # noqa: E501
        :type sb_services: List[SBService]
        :param props: The props of this MECCreation.  # noqa: E501
        :type props: object
        :param geolocation: The geolocation of this MECCreation.  # noqa: E501
        :type geolocation: GeolocationTileList
        """
        self.swagger_types = {
            'name': str,
            'lat': str,
            'lng': str,
            'organization': str,
            'resources': MECCreationResources,
            'sb_services': List[SBService],
            'props': object,
            'geolocation': GeolocationTileList
        }

        self.attribute_map = {
            'name': 'name',
            'lat': 'lat',
            'lng': 'lng',
            'organization': 'organization',
            'resources': 'resources',
            'sb_services': 'sb_services',
            'props': 'props',
            'geolocation': 'geolocation'
        }
        self._name = name
        self._lat = lat
        self._lng = lng
        self._organization = organization
        self._resources = resources
        self._sb_services = sb_services
        self._props = props
        self._geolocation = geolocation

    @classmethod
    def from_dict(cls, dikt) -> 'MECCreation':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The MECCreation of this MECCreation.  # noqa: E501
        :rtype: MECCreation
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this MECCreation.


        :return: The name of this MECCreation.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this MECCreation.


        :param name: The name of this MECCreation.
        :type name: str
        """

        self._name = name

    @property
    def lat(self) -> str:
        """Gets the lat of this MECCreation.


        :return: The lat of this MECCreation.
        :rtype: str
        """
        return self._lat

    @lat.setter
    def lat(self, lat: str):
        """Sets the lat of this MECCreation.


        :param lat: The lat of this MECCreation.
        :type lat: str
        """

        self._lat = lat

    @property
    def lng(self) -> str:
        """Gets the lng of this MECCreation.


        :return: The lng of this MECCreation.
        :rtype: str
        """
        return self._lng

    @lng.setter
    def lng(self, lng: str):
        """Sets the lng of this MECCreation.


        :param lng: The lng of this MECCreation.
        :type lng: str
        """

        self._lng = lng

    @property
    def organization(self) -> str:
        """Gets the organization of this MECCreation.


        :return: The organization of this MECCreation.
        :rtype: str
        """
        return self._organization

    @organization.setter
    def organization(self, organization: str):
        """Sets the organization of this MECCreation.


        :param organization: The organization of this MECCreation.
        :type organization: str
        """

        self._organization = organization

    @property
    def resources(self) -> MECCreationResources:
        """Gets the resources of this MECCreation.


        :return: The resources of this MECCreation.
        :rtype: MECCreationResources
        """
        return self._resources

    @resources.setter
    def resources(self, resources: MECCreationResources):
        """Sets the resources of this MECCreation.


        :param resources: The resources of this MECCreation.
        :type resources: MECCreationResources
        """

        self._resources = resources

    @property
    def sb_services(self) -> List[SBService]:
        """Gets the sb_services of this MECCreation.


        :return: The sb_services of this MECCreation.
        :rtype: List[SBService]
        """
        return self._sb_services

    @sb_services.setter
    def sb_services(self, sb_services: List[SBService]):
        """Sets the sb_services of this MECCreation.


        :param sb_services: The sb_services of this MECCreation.
        :type sb_services: List[SBService]
        """

        self._sb_services = sb_services

    @property
    def props(self) -> object:
        """Gets the props of this MECCreation.


        :return: The props of this MECCreation.
        :rtype: object
        """
        return self._props

    @props.setter
    def props(self, props: object):
        """Sets the props of this MECCreation.


        :param props: The props of this MECCreation.
        :type props: object
        """

        self._props = props

    @property
    def geolocation(self) -> GeolocationTileList:
        """Gets the geolocation of this MECCreation.


        :return: The geolocation of this MECCreation.
        :rtype: GeolocationTileList
        """
        return self._geolocation

    @geolocation.setter
    def geolocation(self, geolocation: GeolocationTileList):
        """Sets the geolocation of this MECCreation.


        :param geolocation: The geolocation of this MECCreation.
        :type geolocation: GeolocationTileList
        """

        self._geolocation = geolocation