# src/apiserver/openapi_server/openapi/cloudinstance.yaml
openapi: 3.0.4
info:
  title: 5G-META Cloud Platform API Server
  description: "API to manage pipeline instances and instace types in different 5GMETA\
    \ MEC Servers. \nCloud components. The Instance API has the scope to consent the\
    \ request of a \npipeline instance and receive the confirmation of the instance\
    \ deployment. \nAn instance can be deleted, but not be modified or updated.\""
  contact:
    name: 5GMETA
    url: https://5gmeta-project.eu/"
    email: 5gmeta@akkodis.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
- url: /api/v1
...
paths:
...
  /mec/locations:
    get:
      tags:
      - Southbound services
      summary: Get serving MEC servers with their locations
      operationId: get_mec_locations
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MECLocationsList'
        "404":
          description: No services found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
  /mec:
    get:
      tags:
      - Southbound services
      summary: Get all serving MEC servers
      operationId: get_mecs
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MECInstanceList'
        "404":
          description: No services found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
    post:
      tags:
      - Southbound services
      summary: Register a MEC instance in the discovery service
      operationId: post_mec
      requestBody:
        description: MEC information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MECCreation'
        required: true
      responses:
        "200":
          description: Successful operation
          content: {}
        "404":
          description: No services found
          content: {}
        "405":
          description: Invalid input
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
  /mec/{mec_id}:
    get:
      tags:
      - Southbound services
      summary: Get a serving MEC server
      operationId: get_mec
      parameters:
      - name: mec_id
        in: path
        description: MEC ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MECInstanceList'
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: No services found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
    delete:
      tags:
      - Southbound services
      summary: Delete a MEC
      operationId: delete_mec
      parameters:
      - name: mec_id
        in: path
        description: MEC ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content: {}
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: No services found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
  /mec/tile:
    get:
      tags:
      - Southbound services
      summary: Get list of tiles
      operationId: get_tiles
      responses:
        "200":
          description: Successful operation
        "401":
          description: Invalid tile supplied
          content: {}
        "404":
          description: No services found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
  /mec/tile/{tile}:
    get:
      tags:
      - Southbound services
      summary: Find serving MEC in a tile
      operationId: get_mec_tile
      parameters:
      - name: tile
        in: path
        description: Tile to get serving MEC
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MECInstanceList'
        "401":
          description: Invalid tile supplied
          content: {}
        "404":
          description: No services found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
  /mec/{mec_id}/tile/{tile}:
    post:
      tags:
      - Southbound services
      summary: Add serving tile to a MEC intance
      operationId: post_tile_mec
      parameters:
      - name: mec_id
        in: path
        description: MEC ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: tile
        in: path
        description: Tile to post
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content: {}
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: No services found
          content: {}
        "405":
          description: Invalid input
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
    delete:
      tags:
      - Southbound services
      summary: Delete a serving tile from a MEC instance
      operationId: delete_tile_mec
      parameters:
      - name: mec_id
        in: path
        description: MEC ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: tile
        in: path
        description: Tile to delete
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content: {}
        "400":
          description: Invalid ID supplied
          content: {}
        "401":
          description: Invalid tile supplied
          content: {}
        "404":
          description: No services found
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
  /mec/{mec_id}/nbservices:
    get:
      tags:
      - Northbound services
      description: Get a northbound service information from a MEC
      operationId: get_nbservices_from_mec
      parameters:
      - name: mec_id
        in: path
        description: MEC ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NBService'
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
    post:
      tags:
      - Northbound services
      description: Add a northbound service information to a MEC
      operationId: add_nbservice_to_mec
      parameters:
      - name: mec_id
        in: path
        description: MEC ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NBService'
      responses:
        "200":
          description: Successful operation
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
  /mec/{mec_id}/nbservices/{service_id}:
    get:
      tags:
      - Northbound services
      description: Get a northbound service from a MEC
      operationId: get_nbservice_from_mec
      parameters:
      - name: mec_id
        in: path
        description: MEC ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: service_id
        in: path
        description: Service ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NBService'
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
    delete:
      tags:
      - Northbound services
      description: Delete a northbound service in a MEC
      operationId: delete_nbservice_in_mec
      parameters:
      - name: mec_id
        in: path
        description: MEC ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: service_id
        in: path
        description: Service ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
    patch:
      tags:
      - Northbound services
      description: Modify a northbound service in a MEC
      operationId: modify_nbservice_in_mec
      parameters:
      - name: mec_id
        in: path
        description: MEC ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: service_id
        in: path
        description: Service ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NBService'
      responses:
        "200":
          description: Successful operation
          content: {}
      x-openapi-router-controller: openapi_server.controllers.mec_services_controller
...
components:
  schemas:
...
    MECCreation:
      type: object
      properties:
        name:
          type: string
        lat:
          type: string
        lng:
          type: string
        organization:
          type: string
        resources:
          type: object
          properties:
            memory:
              type: string
            cpu:
              type: string
            gpu:
              type: string
            storage:
              type: string
        sb_services:
          type: array
          items:
            $ref: '#/components/schemas/SBService'
        props:
          type: object
        geolocation:
          $ref: '#/components/schemas/GeolocationTileList'
      example:
        name: SanSebastian-MEC
        lat: "43.2922071"
        lng: "-1.987018,17"
        organization: Vicomtech
        resources:
          memory: "16"
          cpu: "12"
          gpu: "true"
          storage: 1000"
        sb_services:
        - service_name: message-broker
          description: Message Broker
          host: "<host>"
          port: "5673"
        - service_name: video-broker
          description: Video Stream Broker
          host: "<host>"
          port: "8443"
        - service_name: registration-api
          description: Registration API
          host: "<host>"
          port: "12346"
        props: {}
        geolocation:
        - tile-id: "0331110121011323"
        - tile-id: "0331110121023020"
      xml:
        name: MECInstance
    MECInstance:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        lat:
          type: string
        lng:
          type: string
        organization:
          type: string
        resources:
          type: object
          properties:
            memory:
              type: string
            cpu:
              type: string
            gpu:
              type: string
            storage:
              type: string
        sb_services:
          type: array
          items:
            $ref: '#/components/schemas/SBService'
        props:
          type: object
        geolocation:
          type: array
          items:
            $ref: '#/components/schemas/GeolocationTileList'
      example:
        id: "1"
        name: MEC0
        organization: Organisation
        lat: "43.2922071"
        lng: "-1.987018,17"
        resources:
          memory: "16"
          cpu: "12"
          gpu: "true"
          storage: "1000"
        sb_services:
        - service_name: message-broker
          description: Message Broker
          host: "<host>"
          port: "5673"
        - service_name: video-broker
          description: Video Stream Broker
          host: "<host>"
          port: "8443"
        - service_name: registration-api
          description: Registration API
          host: "<host>"
          port: "12346"
        props: "{}"
        geolocation:
        - tile-id: "033111012101132312"
        - tile-id: "033111012102302012"
      xml:
        name: MECInstance
    MECLocation:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        lat:
          type: string
        lng:
          type: string
        organization:
          type: string
        geolocation:
          type: array
          items:
            $ref: '#/components/schemas/GeolocationTileList'
      example:
        id: "1"
        name: SanSebastian-MEC
        organization: Vicomtech
        geolocation:
        - tile-id: "033111012101132312"
        - tile-id: "033111012102302012"
      xml:
        name: MECInstance
    MECLocationsList:
      type: array
      items:
        $ref: '#/components/schemas/MECLocation'
    MECInstanceList:
      type: array
      items:
        $ref: '#/components/schemas/MECInstance'
    SBService:
      title: SBService
      required:
      - description
      - host
      - port
      - service_name
      type: object
      properties:
        service_name:
          type: string
          description: Name of the exposed service in the MEC
        host:
          type: string
          description: Hostname of the service
          format: hostname
        port:
          type: string
          description: Port of the service
        description:
          type: string
          description: Description of the service
        props:
          type: string
          description: JSON with extra information of the service
          format: json
      example:
        service_name: message-broker
        description: Message Broker
        host: "<host>"
        port: "5673"
        props: props
    NBService:
      title: NBService
      required:
      - host
      - port
      - service_name
      type: object
      properties:
        service_name:
          type: string
          description: Name of the exposed service in the MEC
        host:
          type: string
          description: Hostname of the service
          format: hostname
        port:
          type: string
          description: Port of the service
        description:
          type: string
          description: Description of the service
        props:
          type: string
          description: JSON with extra information of the service
          format: json
      example:
        service_name: message-broker
        description: Message Broker
        host: "<host>"
        port: "61616"
        props: "{}"
    GeolocationTileList:
      type: array
      items:
        $ref: '#/components/schemas/GeolocationTile'
    GeolocationTile:
      type: object
      properties:
        zoom-level:
          type: integer
          default: 18
        tile-id:
          type: string
      xml:
        name: Geolocation
      example:
        tile-id: "033111012101132313"
        zoom-level: 18
...